Set desired yaw to goal yaw inside radius
This commit sets the yaw setpoint to the desired one
once the drone is inside the acceptance radius of the waypoint.
This is necessary for the navigator to check off the mission item.

Note: During normal missions this is not needed, because the desired
yaw that the navigator checks for is the one defined by the parameter
MIS_YAWMODE, which defaults to 1 (toward next mission item). But
when the avoidance is active, it might not arrive at the waypoint
from this direction.