Add script to install commit hooks
This commit adds a script that can be run and will install a pre-commit hook
which runs clang-format on all committed files.

To install the hooks, run (from the main directory)
`./tools/set_up_commit_hooks`

And you'll be set.

Along with it, an addition to the README that explains this.