Remember points based on current point cloud
This commit changes the 'forgetting' of points based on FOV such
that it takes the max and min horizontal component of the current
point cloud to decide. This fixes a bug observed in real hardware,
when a blind spot' at the edge of the FOV due to inaccurate FOV
calculation led to not remembering any points at all.