Plan to sensor horizon at all times
This commit removes the 'max_tree_length' parameter, replacing its
logic with the box_size. This way we always plan to the maximum
of the sensor range, which is a logical thing to do.

To throttle the planning effort the number of nodes expanded during
A* search is much more suitable, and that logic remains untouched.