Make FOV detection robust to cam conventions
This commit changes the way the field of view is calculated dynamically.
We no longer assume the CV convention for the camera (as it doesn't hold
true for some cameras, such as the structure core). Instead we compute
the extrema in the point cloud, transform them into the FCU and compute
the FOV from there.