Compute FOV in camera frame (#420)
* Compute FOV in camera frame

This moves the dynamic calculation of the FOV into the camera frame, such that it is not affected by faulty transforms. Furthermore it is a refactor of the FOV logic, allowing multi-camera FOV to be handled correctly, accounting for discontinous and overlapping fields of view.

This removes the need to subscribe to the camera_info topic.